api/requirements.txt
aniso8601==7.0.0
argon2-cffi==19.2.0
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
graphene==2.1.8
graphene-mongo==0.2.12
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
passlib==1.7.2
promise==2.3
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.12.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
